# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "-",
        "title": "Documenting Whitebook API",
        "contact": {},
        "version": "1"
    },
    "basePath": "/",
    "paths": {
...
        "/book-user-detail/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-user-detail/delete/{bookUserDetailId}": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-user-detail/update": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-user-detail/{bookUserDetailId}": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book-user-details": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book/create": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "-",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
        "/book/update": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "-",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/book/{bookId}": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "-",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "get string by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of books in the the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}